WIP re-enable "-Werror=old-style-cast" (drake_cc.bzl). volumeMetricMesh.cpp has too many pointer castings to fix manually.

+++ tmp/basicAlgorithms.h
@@ -44,7 +44,7 @@
 namespace vegafem
 {
 template<class Container>
-int sizei(const Container & c) { return (int)c.size(); }
+int sizei(const Container & c) { return c.size(); }
 
 // return true if all elements in [first, last] == value
 template<class InputIt, class T>
